splitting install - adding mesh support
